requirements.txt
streamlit
requests
pyyaml
python-dotenv
ibm-watson
ibm-cloud-sdk-core
ibmcloudant
pandas
